refactor(tcf/services): repo folder renamed to tcf
