+++ components/tcf/services/src/infrastructure/bootstrap/ServiceInitializer.js
@@ -6,7 +6,7 @@ import {
   uiVisibleUseCaseFactory,
   updateUserConsentUseCaseFactory
 } from '../../application/service/factory'
-import {tcfRepositoryFactory} from '../tcfx/factory'
+import {tcfRepositoryFactory} from '../tcf/factory'
 
 class ServiceInitializer {
   static init({language = 'es'} = {}) {
